Move ValidateProblemDetails to extension method
